Code cleanup of legacy parser + restructure of directories
