+++ src/parser/mod.rs
@@ -1,0 +1,2 @@
+pub mod osu_parser;
+pub mod osu_util;
