+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 // Utility functions and structs for osu2itg
 pub mod common;
+pub mod constants;
 pub mod file;
